Rework snapshots.
