+++ crates/cli/src/commands/project.rs
@@ -2,6 +2,7 @@ use crate::terminal::{color, ExtendedTerm, Label};
 use console::Term;
 use itertools::Itertools;
 use moon_workspace::Workspace;
+use std::env;
 
 pub async fn project(id: &str, json: &bool) -> Result<(), Box<dyn std::error::Error>> {
     let workspace = Workspace::load()?;
@@ -19,7 +20,11 @@ pub async fn project(id: &str, json: &bool) -> Result<(), Box<dyn std::error::Er
     term.render_label(Label::Brand, &project.id)?;
     term.render_entry("ID", &color::id(&project.id))?;
     term.render_entry("Path", &color::path(&project.location))?;
-    term.render_entry("Root", &color::file_path(&project.dir))?;
+
+    // Dont show in test snapshots
+    if env::var("MOON_TEST").is_err() {
+        term.render_entry("Root", &color::file_path(&project.dir))?;
+    }
 
     if let Some(config) = project.config {
         if let Some(meta) = config.project {
@@ -90,4 +95,85 @@ pub async fn project(id: &str, json: &bool) -> Result<(), Box<dyn std::error::Er
     term.flush()?;
 
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::helpers::{create_test_command, get_assert_output, get_assert_stderr_output};
+    use insta::assert_snapshot;
+
+    #[test]
+    fn unknown_project() {
+        let assert = create_test_command("projects")
+            .arg("project")
+            .arg("unknown")
+            .assert();
+
+        assert_snapshot!(get_assert_stderr_output(&assert));
+
+        assert.failure().code(1);
+    }
+
+    #[test]
+    fn empty_config() {
+        let assert = create_test_command("projects")
+            .arg("project")
+            .arg("emptyConfig")
+            .assert();
+
+        assert_snapshot!(get_assert_output(&assert));
+    }
+
+    #[test]
+    fn no_config() {
+        let assert = create_test_command("projects")
+            .arg("project")
+            .arg("noConfig")
+            .assert();
+
+        assert_snapshot!(get_assert_output(&assert));
+    }
+
+    #[test]
+    fn basic_config() {
+        // with dependsOn and fileGroups
+        let assert = create_test_command("projects")
+            .arg("project")
+            .arg("basic")
+            .assert();
+
+        assert_snapshot!(get_assert_output(&assert));
+    }
+
+    #[test]
+    fn advanced_config() {
+        // with project metadata
+        let assert = create_test_command("projects")
+            .arg("project")
+            .arg("advanced")
+            .assert();
+
+        assert_snapshot!(get_assert_output(&assert));
+    }
+
+    #[test]
+    fn depends_on_paths() {
+        // shows dependsOn paths when they exist
+        let assert = create_test_command("projects")
+            .arg("project")
+            .arg("foo")
+            .assert();
+
+        assert_snapshot!(get_assert_output(&assert));
+    }
+
+    #[test]
+    fn with_tasks() {
+        let assert = create_test_command("projects")
+            .arg("project")
+            .arg("tasks")
+            .assert();
+
+        assert_snapshot!(get_assert_output(&assert));
+    }
 }
